# pkg/api/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
            "engine.CreateNetworkParams": {
                "properties": {
                    "address": {
                        "type": "string"
                    },
                    "device": {
                        "type": "string"
                    },
                    "members": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "mode": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "parent": {
                        "type": "string"
                    },
                    "tag": {
                        "type": "integer"
                    },
                    "uplink": {
                        "type": "string"
                    },
                    "vlans": {
                        "items": {
                            "$ref": "#/components/schemas/types.VLAN"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "jobs.Job": {
                "properties": {
                    "action": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    },
                    "finished_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    },
                    "logs": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "result": {
                        "type": "string"
                    },
                    "started_at": {
                        "type": "string"
                    },
                    "state": {
                        "$ref": "#/components/schemas/jobs.State"
                    },
                    "target": {
                        "type": "string"
                    }
                },
                "required": [
                    "action",
                    "created_at",
                    "id",
                    "label",
                    "logs",
                    "state",
                    "target"
                ],
                "type": "object"
            },
...
            "jobs.State": {
                "enum": [
                    "pending",
                    "running",
                    "succeeded",
                    "failed"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "Pending",
                    "Running",
                    "Succeeded",
                    "Failed"
                ]
            },
...
            "types.NetworkManifest": {
                "properties": {
                    "address": {
                        "type": "string"
                    },
                    "applied_address": {
                        "type": "string"
                    },
                    "applied_members": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "applied_vlans": {
                        "items": {
                            "$ref": "#/components/schemas/types.VLAN"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "device": {
                        "type": "string"
                    },
                    "group": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "members": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "mode": {
                        "$ref": "#/components/schemas/types.NetworkMode"
                    },
                    "name": {
                        "type": "string"
                    },
                    "owned": {
                        "type": "boolean"
                    },
                    "parent": {
                        "type": "string"
                    },
                    "tag": {
                        "type": "integer"
                    },
                    "uplink": {
                        "type": "string"
                    },
                    "vlans": {
                        "items": {
                            "$ref": "#/components/schemas/types.VLAN"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "required": [
                    "id",
                    "mode",
                    "name"
                ],
                "type": "object"
            },
            "types.NetworkMode": {
                "enum": [
                    "user",
                    "vmnet-bridged",
                    "bridge",
                    "switch",
                    "bridged",
                    "vlan",
                    "bridge",
                    "switch",
                    "bridged",
                    "vmnet-bridged",
                    "user",
                    "vlan"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "NetworkUser",
                    "NetworkVmnetBridged",
                    "NetworkBridge",
                    "NetworkSwitch",
                    "NetworkBridged",
                    "NetworkVLAN"
                ]
            },
            "types.VLAN": {
                "properties": {
                    "borrowed": {
                        "type": "boolean"
                    },
                    "device": {
                        "type": "string"
                    },
                    "parent": {
                        "type": "string"
                    },
                    "tag": {
                        "maximum": 4094,
                        "minimum": 1,
                        "type": "integer"
                    }
                },
                "required": [
                    "parent",
                    "tag"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer JWT obtained from POST /api/login.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "VM lifecycle, networking, media and jobs API. Mutations returning 202 are asynchronous; inspect the returned job for completion.",
        "title": "Maco API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/networks": {
            "get": {
                "operationId": "listNetworks",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/types.NetworkManifest"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "listNetworks",
                "tags": [
                    "networks"
                ]
            },
            "post": {
                "operationId": "createNetwork",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/engine.CreateNetworkParams",
                                        "summary": "body",
                                        "description": "Request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "createNetwork",
                "tags": [
                    "networks"
                ]
            }
        },
        "/api/networks/{id}": {
            "delete": {
                "operationId": "destroyNetwork",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "destroyNetwork",
                "tags": [
                    "networks"
                ]
            },
            "put": {
                "operationId": "updateNetwork",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/engine.CreateNetworkParams",
                                        "summary": "body",
                                        "description": "Request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "updateNetwork",
                "tags": [
                    "networks"
                ]
            }
        },
        "/api/networks/{id}/apply": {
            "post": {
                "operationId": "applyNetwork",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "applyNetwork",
                "tags": [
                    "networks"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ]
}